backend/requirements.txt
accessify==0.3.1
asgiref==3.5.2
Django==3.2.16
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.4.0
gunicorn==20.0.4
djangorestframework==3.14.0
django-filter==22.1
djangorestframework-simplejwt==4.8.0
drf-extra-fields==3.4.1
djoser==2.1.0
Pillow==9.3.0
django-import-export==3.0.2
psycopg2-binary==2.8.6
xhtml2pdf==0.2.8
